moved multiagent environment to a new folder
