+++ flow/envs/loop/loop_accel.py
@@ -2,7 +2,6 @@
 
 from flow.core import rewards
 from flow.envs.base_env import Env
-from flow.envs.multiagent_env import MultiEnv
 
 from gym.spaces.box import Box
 
@@ -105,49 +104,3 @@ class AccelEnv(Env):
         if self.vehicles.num_rl_vehicles > 0:
             for veh_id in self.vehicles.get_human_ids():
                 self.vehicles.set_observed(veh_id)
-
-
-class MultiAgentAccelEnv(AccelEnv, MultiEnv):
-    """Adversarial multi-agent env.
-
-    Multi-agent env with an adversarial agent perturbing
-    the accelerations of the autonomous vehicle
-    """
-    def _apply_rl_actions(self, rl_actions):
-        """See class definition."""
-        sorted_rl_ids = [
-            veh_id for veh_id in self.sorted_ids
-            if veh_id in self.vehicles.get_rl_ids()
-        ]
-        av_action = rl_actions['av']
-        adv_action = rl_actions['adversary']
-        perturb_weight = self.env_params.additional_params['perturb_weight']
-        rl_action = av_action + perturb_weight * adv_action
-        self.apply_acceleration(sorted_rl_ids, rl_action)
-
-    def compute_reward(self, rl_actions, **kwargs):
-        """The agents receives opposing speed rewards.
-
-        The agent receives the class definition reward,
-        the adversary receives the negative of the agent reward
-        """
-        if self.env_params.evaluate:
-            reward = np.mean(self.vehicles.get_speed(self.vehicles.get_ids()))
-            return {'av': reward, 'adversary': -reward}
-        else:
-            reward = rewards.desired_velocity(self, fail=kwargs['fail'])
-            return {'av': reward, 'adversary': -reward}
-
-    def get_state(self, **kwargs):
-        """See class definition for the state.
-
-        The adversary state and the agent state are identical.
-        """
-        # speed normalizer
-        max_speed = self.scenario.max_speed
-        state = np.array([[
-            self.vehicles.get_speed(veh_id) / max_speed,
-            self.get_x_by_id(veh_id) / self.scenario.length
-        ] for veh_id in self.sorted_ids])
-        state = np.ndarray.flatten(state)
-        return {'av': state, 'adversary': state}
