+++ flow/visualize/visualizer_rllib.py
@@ -23,6 +23,7 @@ from ray.rllib.agents.agent import get_agent_class
 from ray.tune.registry import register_env
 from ray.rllib.models import ModelCatalog
 
+import flow.envs
 from flow.core.util import emission_to_csv
 from flow.utils.registry import make_create_env
 from flow.utils.rllib import get_flow_params
@@ -132,9 +133,19 @@ def visualizer_rllib(args):
         net_params=net_params,
         initial_config=initial_config)
 
+    # check if the environment is a single or multiagent environment, and
+    # get the right address accordingly
+    single_agent_envs = [env for env in dir(flow.envs)
+                         if not env.startswith('__')]
+
+    if flow_params['env_name'] in single_agent_envs:
+        env_loc = 'flow.envs'
+    else:
+        env_loc = 'flow.multiagent_envs'
+
     # Start the environment with the gui turned on and a path for the
     # emission file
-    module = __import__('flow.envs', fromlist=[flow_params['env_name']])
+    module = __import__(env_loc, fromlist=[flow_params['env_name']])
     env_class = getattr(module, flow_params['env_name'])
     env_params = flow_params['env']
     env_params.restart_instance = False
